Factoring benchmarking methods.
